new file tag and some functions are moved
