+++ pkg/bit/v2/decode.go
@@ -22,51 +22,9 @@ import (
 	"fmt"
 	"github.com/goccy/go-reflect"
 	"io"
-	"strings"
-)
-
-const (
-	tagKeyName = "bit"
 )
 
 var errCannotInterface = errors.New("CanInterface returns false")
-
-// tagConfig represents StructTag.
-//   "-"   : ignore the field
-//   "skip": ignore but offset will be updated
-//   "BE"  : the field is treated as big endian
-//   "LE"  : the field is treated as little endian
-type tagConfig struct {
-	ignore bool
-	skip   bool
-	endian binary.ByteOrder
-}
-
-func parseStructTag(t reflect.StructTag) *tagConfig {
-	s, ok := t.Lookup(tagKeyName)
-	if !ok {
-		return nil
-	}
-	ret := &tagConfig{}
-
-	strs := strings.Split(s, ",")
-	for _, v := range strs {
-		switch v {
-		case "-":
-			ret.ignore = true
-			return ret
-		case "skip":
-			ret.skip = true
-			return ret
-		case "BE":
-			ret.endian = binary.BigEndian
-		case "LE":
-			ret.endian = binary.LittleEndian
-		}
-
-	}
-	return ret
-}
 
 // Size returns size of v in bits.
 /*
